requirements.txt
customtkinter
cryptography
requests
pillow
selenium
webdriver_manager
iitkgp-erp-login
